동적 메모리 할당 추가

+++ src/kernel64/util/memory.h
@@ -6,6 +6,6 @@
 void kMemSet(void* pvDestination, BYTE bData, int iSize);
 int kMemCpy(void* pvDestination, const void* pvSource, int iSize);
 int kMemCmp(const void* pvDestination, const void* pvSource, int iSize);
-int kMemSize();
+QWORD kMemSize();
 
 #endif /*__UTILITY_H__*/
